Move the development database settings to settings.development
This means settings.base does not define a database anymore, and each
site-specific settings file should define it's own. For now, the
settings.production does not define one. This must be done when including this
change in the deployment branch, and will probably mean reading the database
settings from a configuration file.