Install postgres and create a database during the initial setup
We set a random(?) password for the pypln database user and store it in
`~/.pgpass`[1]. The production settings file should read this and use it. That
way the password will be unique for each deploy environment and won't be under
version control. The way we create the random password could probably be
improved.

We still need to use the production settings file to run django.

[1] http://www.postgresql.org/docs/8.1/static/libpq-pgpass.html