Stops `git pulling` when updating code
Instead, we now reset the branch to be the same as origin (creating the branch
if it doesn't exist. Git pull risked creating merge commits (and thus a
diverging tree), which could complicate things. The production host should
always have the working copy exactly the same as the upstream repository.