Upgrade virtualenv during initial_setup
Virtualenv's default behaviour has changed from allowing access to system site
packages to not allowing it, and the command line option to enforce the "old"
behaviour (--system-site-packages) is new. Updating virtualenv not only gives
us newer (and hopefully less buggy) software, but also ensures we know the
behaviour we'll get with regards to the inclusion of system site packages in
the path.