Runs the application in wsgi (with gunicorn)
Closes #78

Note: This introduces a weird issue in my test environment. I'm redirecting my
local port 9000 to port 80 on the host that's running pypln. This has worked
fine until now but with this configuration the redirects after POSTs lead the
browser to port 80 instead. This looks like a small issue for now, but we
should look into it and see if there isn't a more serious symptom.