Fixes error for files with non ascii chars encoded in utf-8
We should probably try to work with unicode for as long as possible. Decode the
content right after receiving it, and enconding it only just before returning
it. And this brings us to the problem with this code: it assumes that
file_data['contents'] is utf-8. It looks like GridFS (which we use in
pypln.web) stores content in the encoding the user sends, so it breaks this
code. Please don't use this until we solve this issue.