Uses both the 7 classes and the 3 classes classifier in the NER worker
Since the 3 classes classifier is more acurate, it is better to extract the
classes it recognizes. The rest of the classes are extracted using the 7
classes worker. One thing to note: this means that the 'O' class will have
entities that were recognized by the 7 classes classifier, but not the 3
classes.