Adds dependencies for wordcloud generation
There is a problem when installing the wordcloud module if Cython was not
previously installed. If that happens, installing Cython before trying to
install all deps (with `pip install -r`) should solve this.

I forked the original wordcloud module (https://github.com/amueller/word_cloud)
because it needed to be slightly adapted (so it wouldn't show and save the
image, for example).