Fixes `UnicodeDecodeError` in `PalavrasRaw` worker
For some reason, in some pypln installations the text extracted from the
document was not getting to the `PalavrasRaw` worker as unicode. This may be
due to previous errors during the decoding process that we fixed earlier. That
meant that, when we got a non-unicode string, python would try to decode it
using the default codec (ascii) in `text.encode(PALAVRAS_ENCODING)`. Since we
know the text came from mongodb, we can just decode it using utf-8 to make sure
we have a unicode object.