Makes sure we use the correct codec in all Palavras workers
In some instalations (in our production server, for example)
`sys.getfilesystemencoding()` was not returning the correct encoding when the
worker was run by pypelinin's broker. This meant that, when sending text to
palavras stdin, we were using the wrong encoding, resulting in a
`UnicodeEncodeError`. This commits forces utf-8 as the encoding to be used when
communicating with the process and also adds some code to make sure we use the
correct encoding everywhere in the workers that depend on palavras.