WIP: starts to change Extractor worker into a Celery task
The tests are broken here. For some reason, when running the tests, Celery
can't find the correct Celery app (the one in `pypln.backend.celery_app:app`
and uses the default one, which means that it doesn't have a result backend
configured. I've commit this work in progress so others can take a look and see
what I'm doing wrong here.