Fixes and documents the issue with the app import
The comment above the new import in `pypln/backend/celery_task.py`[1] explains
the issue and why this import is necessary. Thank you hidden collateral effects
for taking 2 days out of my life.

[1]: "This import may look like an unused imported, but it is not.
When our base task class is defined, the Celery app must have already been
instantiated, otherwise when this code is imported elsewhere (like in a
client that will call a task, for example) celery will fallback to the
default app, and our configuration will be ignored. This is not an issue in
the documented project layout, because there they import the app in the
module that define the tasks (to use the decorator in `app.task`)."