Adds Task to retrieve filedata from GridFS
This was previously done as a special case in `MongoDBStore` (when the worker
that asked for information was `Extractor` it would fetch data from gridfs).
But this is not very different from other Tasks in the sense that it needs some
information (the ObjectId for the file in GridFS) and populates the document
with information needed by other workers (such as the file contents for the
Extractor) so I decided to move this to it's own task (instead of including
this logic in the Extractor).