Makes sure tests only run if the database name starts with `test`
The idea is to make sure we don't drop the production database if we run tests.
The best way to implement this is probably a test runner that adds a `test`
prefix to whichever database name is defined, but we need to make sure tasks
will also see the changed configuration.