Changes GridFSDataRetriever to base64 encode the contents before storing
This is necessary because the content of the file might be binary data (such as
a pdf), and we don't want to change it by encoding/decoding. This was not an
issue before because MongoDict used to pickle this before storing. This kind of
adaptation is probably going to happen in other places where the mapping from
python objects to json is not straight forward.