Adapts Trigrams worker to the removal of MongoDict
This commit was co-authored by Israel Teixeira <israelst@gmail.com>

We needed to store the trigram results as strings (since that's the only
possible type for mongo keys). We decided to turn the tuples into strings
joined by spaces because spaces are never going to be part of a token.

Also, mongo keys cannot contain neither `.` nor `$`, so we decided to replace
those with `\dot` and `\dollarsign` respectively.

Thanks a lot @israelst for the help!